+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,8 @@
 package racingcar;
 
+import racingcar.application.Game;
+import racingcar.common.GameConfig;
+
 public class Application {
     public static void main(String[] args) {
         GameConfig gameConfig = GameConfig.getInstance();
