+++ src/main/java/racingcar/application/Game.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.application;
 
 public interface Game {
