+++ src/main/java/racingcar/application/RacingManager.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.application;
 
 import java.util.List;
+import racingcar.racing.Racer;
 
 public interface RacingManager<T extends Racer> {
