+++ src/main/java/racingcar/application/service/RacingCarGame.java
@@ -1,6 +1,11 @@
-package racingcar;
+package racingcar.application.service;
 
 import java.util.List;
+import racingcar.RacingCar;
+import racingcar.application.Converter;
+import racingcar.application.Game;
+import racingcar.application.RacingManager;
+import racingcar.view.ApplicationView;
 
 public class RacingCarGame implements Game {
