+++ src/main/java/racingcar/common/ConsoleMessage.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.common;
 
 public final class ConsoleMessage {
