+++ src/main/java/racingcar/racing/Racer.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.racing;
 
 public abstract class Racer {
