+++ src/main/java/racingcar/view/ApplicationView.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 public interface ApplicationView {
