+++ src/test/java/racingcar/RacingCarConverterTest.java
@@ -6,6 +6,7 @@ import camp.nextstep.edu.missionutils.test.Assertions;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.application.service.RacingCarConverter;
 
 class RacingCarConverterTest {
